Split OpportunityDetailsFields into separate components for each field and split the header from NewOpportunity into it's own component
