+++ src/components/Dashboard/NewOpportunity/NewOpportunity.tsx
@@ -1,5 +1,4 @@
 "use client";
-import { EditableField } from "@/components/EditableField/EditableField";
 import { SectionCard } from "@/components/Dashboard/Profile/common/SectionCard";
 import {
   useApiActivities,
@@ -13,58 +12,28 @@ import {
 import { AccompanyingDetailsEdit } from "@/components/Dashboard/Profile/sections/AccompanyingDetails/AccompanyingDetailsEdit";
 import { BackButton, PageContainer } from "@/components/Dashboard/Profile/styles";
 import { IconName } from "@/components/Dashboard/Profile/types";
-import {
-  Card,
-  IconContainer,
-  ProfileContent,
-  ProfileInfo,
-  StatusSection,
-  TitleSection,
-} from "@/components/Dashboard/Profile/sections/ProfileHeader/common/profileHeaderStyles";
-import { createVolunteerTypeLabelMap } from "@/components/Dashboard/Profile/sections/ProfileHeader/common/labelMaps";
 import Button from "@/components/core/button/Button/Button";
 import { apiPathOpportunity, DashboardRoutes } from "@/config/constants";
 import { useMutationQuery } from "@/hooks";
 import { useGetCurrentAgent } from "@/hooks/useGetCurrentAgent";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Heading2, Heading4 } from "@/components/styled/text";
-import { ShootingStarIcon, ArrowLeftIcon } from "@phosphor-icons/react";
+import { Heading2 } from "@/components/styled/text";
+import { ArrowLeftIcon } from "@phosphor-icons/react";
 import { de, enUS } from "date-fns/locale";
 import { TranslatedIntoType, VolunteerStateTypeType, OpportunityFormDataWithAgentSubmitter } from "need4deed-sdk";
 import { useMemo } from "react";
 import { useRouter } from "next/navigation";
-import { Controller, FormProvider, useForm } from "react-hook-form";
+import { FormProvider, useForm } from "react-hook-form";
 import { useTranslation } from "react-i18next";
-import { z } from "zod";
 import {
   createAccompanyingDetailsSchema,
   AccompanyingDetailsFormData,
 } from "../Profile/sections/AccompanyingDetails/createAccompanyingDetailsSchema";
-import { VolunteerTypeRow, TypeButtons, TypeButton, SaveRow } from "./styled";
-import { OpportunityDetailsFields } from "./OpportunityDetailsFields";
+import { SaveRow } from "./styled";
+import { OpportunityDetailsFields } from "./fields/OpportunityDetailsFields";
 import { buildCreatePayload } from "./helper";
-
-// ─── Types ──────────────────────────────────────────────────────────────────
-
-const SELECTABLE_VOLUNTEER_TYPES = [
-  VolunteerStateTypeType.REGULAR,
-  VolunteerStateTypeType.ACCOMPANYING,
-  VolunteerStateTypeType.EVENTS,
-] as const;
-
-const createHeaderSchema = (t: (key: string) => string) =>
-  z.object({
-    title: z.string().min(1, t("form.error.required")),
-    volunteerType: z.enum([
-      VolunteerStateTypeType.REGULAR,
-      VolunteerStateTypeType.ACCOMPANYING,
-      VolunteerStateTypeType.EVENTS,
-    ]),
-  });
-
-export type HeaderFormData = z.infer<ReturnType<typeof createHeaderSchema>>;
-
-// ─── Main component ──────────────────────────────────────────────────────────
+import { createHeaderSchema, HeaderFormData } from "./headerSchema";
+import OpportunityHeaderCard from "./OpportunityHeaderCard";
 
 export function NewOpportunity() {
   const { t, i18n } = useTranslation();
@@ -72,8 +41,6 @@ export function NewOpportunity() {
   const locale = lang === "de" ? de : enUS;
   const router = useRouter();
   const { agentId } = useGetCurrentAgent();
-
-  const volunteerTypeLabelMap = createVolunteerTypeLabelMap(t);
 
   const { data: apiLanguages = [] } = useApiLanguages();
   const { data: apiActivities = [] } = useApiActivities();
@@ -85,11 +52,7 @@ export function NewOpportunity() {
     mode: "onChange",
     defaultValues: { title: "", volunteerType: VolunteerStateTypeType.REGULAR },
   });
-  const {
-    control: headerControl,
-    watch: watchHeader,
-    formState: { errors: headerErrors },
-  } = headerMethods;
+  const { watch: watchHeader } = headerMethods;
   const selectedType = watchHeader("volunteerType");
   const isAccompanying = selectedType === VolunteerStateTypeType.ACCOMPANYING;
   const isEvent = selectedType === VolunteerStateTypeType.EVENTS;
@@ -182,49 +145,9 @@ export function NewOpportunity() {
       <Heading2>{t("dashboard.newOpportunity.title")}</Heading2>
 
       {/* Header card — title input + volunteer type selector */}
-      <Card>
-        <ProfileContent>
-          <IconContainer>
-            <ShootingStarIcon size={120} color="var(--color-blue-500)" weight="duotone" />
-          </IconContainer>
-          <ProfileInfo>
-            <TitleSection>
-              <Controller
-                name="title"
-                control={headerControl}
-                render={({ field }) => (
-                  <EditableField
-                    mode="edit"
-                    type="text"
-                    label={t("dashboard.newOpportunity.fields.title")}
-                    value={field.value}
-                    setValue={field.onChange}
-                    errorMessage={headerErrors.title?.message}
-                  />
-                )}
-              />
-            </TitleSection>
-
-            <StatusSection>
-              <VolunteerTypeRow>
-                <Heading4>{t("dashboard.volunteerProfile.volunteerHeader.volunteerType_title")}</Heading4>
-                <TypeButtons>
-                  {SELECTABLE_VOLUNTEER_TYPES.map((type) => (
-                    <TypeButton
-                      key={type}
-                      type="button"
-                      $active={selectedType === type}
-                      onClick={() => headerMethods.setValue("volunteerType", type, { shouldValidate: true })}
-                    >
-                      {volunteerTypeLabelMap[type]}
-                    </TypeButton>
-                  ))}
-                </TypeButtons>
-              </VolunteerTypeRow>
-            </StatusSection>
-          </ProfileInfo>
-        </ProfileContent>
-      </Card>
+      <FormProvider {...headerMethods}>
+        <OpportunityHeaderCard selectedType={selectedType} />
+      </FormProvider>
 
       {/* Opportunity Details section */}
       <SectionCard
